pyproject.toml
[project]
name = "ddd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "aiomysql>=0.3.2",
    "asyncmy>=0.2.10",
    "configdict>=2.13.1",
    "fastapi>=0.121.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]
